feat: implement product and category management with CRUD operations and SKU support

+++ internal/utils/response.go
@@ -15,10 +15,10 @@ type Response struct {
 }
 
 type PaginationMeta struct {
-	Page       int `json:"page"`
-	Limit      int `json:"limit"`
-	Total      int `json:"total"`
-	TotalPages int `json:"total_pages"`
+	Page       int   `json:"page"`
+	Limit      int   `json:"limit"`
+	Total      int64 `json:"total"`
+	TotalPages int   `json:"total_pages"`
 }
 
 type PaginatedResponse struct {
